// gen/common_pb.ts
export class Location extends Message<Location> {
  /**
   * The file path
   *
   * @generated from field: optional string file = 1;
   */
  file?: string;

  /**
   * The Rust module path
   *
   * @generated from field: optional string module_path = 2;
   */
  modulePath?: string;

  /**
   * The line number in the source code file.
   *
   * @generated from field: optional uint32 line = 3;
   */
  line?: number;

  /**
   * The character in `line`.
   *
   * @generated from field: optional uint32 column = 4;
   */
  column?: number;

  constructor(data?: PartialMessage<Location>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "rs.tokio.console.common.Location";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "file", kind: "scalar", T: 9 /* ScalarType.STRING */, opt: true },
    { no: 2, name: "module_path", kind: "scalar", T: 9 /* ScalarType.STRING */, opt: true },
    { no: 3, name: "line", kind: "scalar", T: 13 /* ScalarType.UINT32 */, opt: true },
    { no: 4, name: "column", kind: "scalar", T: 13 /* ScalarType.UINT32 */, opt: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Location {
    return new Location().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Location {
    return new Location().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Location {
    return new Location().fromJsonString(jsonString, options);
  }

  static equals(a: Location | PlainMessage<Location> | undefined, b: Location | PlainMessage<Location> | undefined): boolean {
    return proto3.util.equals(Location, a, b);
  }
}

/**
 * Unique identifier for metadata.
 *
 * @generated from message rs.tokio.console.common.MetaId
 */
export class MetaId extends Message<MetaId> {
  /**
   * The unique identifier's concrete value.
   *
   * @generated from field: uint64 id = 1;
   */
  id = protoInt64.zero;

  constructor(data?: PartialMessage<MetaId>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "rs.tokio.console.common.MetaId";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "id", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): MetaId {
    return new MetaId().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): MetaId {
    return new MetaId().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): MetaId {
    return new MetaId().fromJsonString(jsonString, options);
  }

  static equals(a: MetaId | PlainMessage<MetaId> | undefined, b: MetaId | PlainMessage<MetaId> | undefined): boolean {
    return proto3.util.equals(MetaId, a, b);
  }
}
// ...
export class Field extends Message<Field> {
  /**
   * The key of the key-value pair.
   *
   * This is either represented as a string, or as an index into a `Metadata`'s 
   * array of field name strings.
   *
   * @generated from oneof rs.tokio.console.common.Field.name
   */
  name: {
    /**
     * The string representation of the name.
     *
     * @generated from field: string str_name = 1;
     */
    value: string;
    case: "strName";
  } | {
    /**
     * An index position into the `Metadata.field_names` of the metadata
     * for the task span that the field came from.
     *
     * @generated from field: uint64 name_idx = 2;
     */
    value: bigint;
    case: "nameIdx";
  } | { case: undefined; value?: undefined } = { case: undefined };

  /**
   * The value of the key-value pair.
   *
   * @generated from oneof rs.tokio.console.common.Field.value
   */
  value: {
    /**
     * A value serialized to a string using `fmt::Debug`.
     *
     * @generated from field: string debug_val = 3;
     */
    value: string;
    case: "debugVal";
  } | {
    /**
     * A string value.
     *
     * @generated from field: string str_val = 4;
     */
    value: string;
    case: "strVal";
  } | {
    /**
     * An unsigned integer value.
     *
     * @generated from field: uint64 u64_val = 5;
     */
    value: bigint;
    case: "u64Val";
  } | {
    /**
     * A signed integer value.
     *
     * @generated from field: sint64 i64_val = 6;
     */
    value: bigint;
    case: "i64Val";
  } | {
    /**
     * A boolean value.
     *
     * @generated from field: bool bool_val = 7;
     */
    value: boolean;
    case: "boolVal";
  } | { case: undefined; value?: undefined } = { case: undefined };

  /**
   * Metadata for the task span that the field came from.
   *
   * @generated from field: rs.tokio.console.common.MetaId metadata_id = 8;
   */
  metadataId?: MetaId;

  constructor(data?: PartialMessage<Field>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "rs.tokio.console.common.Field";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "str_name", kind: "scalar", T: 9 /* ScalarType.STRING */, oneof: "name" },
    { no: 2, name: "name_idx", kind: "scalar", T: 4 /* ScalarType.UINT64 */, oneof: "name" },
    { no: 3, name: "debug_val", kind: "scalar", T: 9 /* ScalarType.STRING */, oneof: "value" },
    { no: 4, name: "str_val", kind: "scalar", T: 9 /* ScalarType.STRING */, oneof: "value" },
    { no: 5, name: "u64_val", kind: "scalar", T: 4 /* ScalarType.UINT64 */, oneof: "value" },
    { no: 6, name: "i64_val", kind: "scalar", T: 18 /* ScalarType.SINT64 */, oneof: "value" },
    { no: 7, name: "bool_val", kind: "scalar", T: 8 /* ScalarType.BOOL */, oneof: "value" },
    { no: 8, name: "metadata_id", kind: "message", T: MetaId },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Field {
    return new Field().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Field {
    return new Field().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Field {
    return new Field().fromJsonString(jsonString, options);
  }

  static equals(a: Field | PlainMessage<Field> | undefined, b: Field | PlainMessage<Field> | undefined): boolean {
    return proto3.util.equals(Field, a, b);
  }
}